requirement.txt
asn1crypto==0.24.0
asyncio==3.4.3
asyncssh==1.16.1
bcrypt==3.1.6
cffi==1.12.2
cryptography==2.6.1
netdev==0.9.0
netmiko==2.3.3
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyserial==3.4
PyYAML==5.1
scp==0.13.2
six==1.12.0
textfsm==0.4.1
